+++ src/cpp11.cpp
@@ -49,10 +49,10 @@ extern "C" SEXP _onnxr_onnx_create_env() {
   END_CPP11
 }
 // session.cpp
-SEXP onnx_create_session(SEXP env_ptr, std::string model_path, std::string provider, std::string cache_dir, int threads, int opt_level, cpp11::strings external_data_files);
-extern "C" SEXP _onnxr_onnx_create_session(SEXP env_ptr, SEXP model_path, SEXP provider, SEXP cache_dir, SEXP threads, SEXP opt_level, SEXP external_data_files) {
+SEXP onnx_create_session(SEXP env_ptr, std::string model_path, std::string provider, std::string cache_dir, int threads, int opt_level);
+extern "C" SEXP _onnxr_onnx_create_session(SEXP env_ptr, SEXP model_path, SEXP provider, SEXP cache_dir, SEXP threads, SEXP opt_level) {
   BEGIN_CPP11
-    return cpp11::as_sexp(onnx_create_session(cpp11::as_cpp<cpp11::decay_t<SEXP>>(env_ptr), cpp11::as_cpp<cpp11::decay_t<std::string>>(model_path), cpp11::as_cpp<cpp11::decay_t<std::string>>(provider), cpp11::as_cpp<cpp11::decay_t<std::string>>(cache_dir), cpp11::as_cpp<cpp11::decay_t<int>>(threads), cpp11::as_cpp<cpp11::decay_t<int>>(opt_level), cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(external_data_files)));
+    return cpp11::as_sexp(onnx_create_session(cpp11::as_cpp<cpp11::decay_t<SEXP>>(env_ptr), cpp11::as_cpp<cpp11::decay_t<std::string>>(model_path), cpp11::as_cpp<cpp11::decay_t<std::string>>(provider), cpp11::as_cpp<cpp11::decay_t<std::string>>(cache_dir), cpp11::as_cpp<cpp11::decay_t<int>>(threads), cpp11::as_cpp<cpp11::decay_t<int>>(opt_level)));
   END_CPP11
 }
 // session.cpp
@@ -105,6 +105,13 @@ extern "C" SEXP _onnxr_onnx_session_input_types(SEXP session_ptr) {
   END_CPP11
 }
 // session.cpp
+cpp11::writable::logicals onnx_session_input_optional(SEXP session_ptr);
+extern "C" SEXP _onnxr_onnx_session_input_optional(SEXP session_ptr) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(onnx_session_input_optional(cpp11::as_cpp<cpp11::decay_t<SEXP>>(session_ptr)));
+  END_CPP11
+}
+// session.cpp
 cpp11::writable::integers onnx_session_output_types(SEXP session_ptr);
 extern "C" SEXP _onnxr_onnx_session_output_types(SEXP session_ptr) {
   BEGIN_CPP11
@@ -114,21 +121,22 @@ extern "C" SEXP _onnxr_onnx_session_output_types(SEXP session_ptr) {
 
 extern "C" {
 static const R_CallMethodDef CallEntries[] = {
-    {"_onnxr_onnx_create_env",            (DL_FUNC) &_onnxr_onnx_create_env,            0},
-    {"_onnxr_onnx_create_session",        (DL_FUNC) &_onnxr_onnx_create_session,        7},
-    {"_onnxr_onnx_is_loaded",             (DL_FUNC) &_onnxr_onnx_is_loaded,             0},
-    {"_onnxr_onnx_load_lib",              (DL_FUNC) &_onnxr_onnx_load_lib,              1},
-    {"_onnxr_onnx_run_",                  (DL_FUNC) &_onnxr_onnx_run_,                  6},
-    {"_onnxr_onnx_session_input_count",   (DL_FUNC) &_onnxr_onnx_session_input_count,   1},
-    {"_onnxr_onnx_session_input_names",   (DL_FUNC) &_onnxr_onnx_session_input_names,   1},
-    {"_onnxr_onnx_session_input_shapes",  (DL_FUNC) &_onnxr_onnx_session_input_shapes,  1},
-    {"_onnxr_onnx_session_input_types",   (DL_FUNC) &_onnxr_onnx_session_input_types,   1},
-    {"_onnxr_onnx_session_output_count",  (DL_FUNC) &_onnxr_onnx_session_output_count,  1},
-    {"_onnxr_onnx_session_output_names",  (DL_FUNC) &_onnxr_onnx_session_output_names,  1},
-    {"_onnxr_onnx_session_output_shapes", (DL_FUNC) &_onnxr_onnx_session_output_shapes, 1},
-    {"_onnxr_onnx_session_output_types",  (DL_FUNC) &_onnxr_onnx_session_output_types,  1},
-    {"_onnxr_onnx_unload_lib",            (DL_FUNC) &_onnxr_onnx_unload_lib,            0},
-    {"_onnxr_onnx_version",               (DL_FUNC) &_onnxr_onnx_version,               0},
+    {"_onnxr_onnx_create_env",             (DL_FUNC) &_onnxr_onnx_create_env,             0},
+    {"_onnxr_onnx_create_session",         (DL_FUNC) &_onnxr_onnx_create_session,         6},
+    {"_onnxr_onnx_is_loaded",              (DL_FUNC) &_onnxr_onnx_is_loaded,              0},
+    {"_onnxr_onnx_load_lib",               (DL_FUNC) &_onnxr_onnx_load_lib,               1},
+    {"_onnxr_onnx_run_",                   (DL_FUNC) &_onnxr_onnx_run_,                   6},
+    {"_onnxr_onnx_session_input_count",    (DL_FUNC) &_onnxr_onnx_session_input_count,    1},
+    {"_onnxr_onnx_session_input_names",    (DL_FUNC) &_onnxr_onnx_session_input_names,    1},
+    {"_onnxr_onnx_session_input_optional", (DL_FUNC) &_onnxr_onnx_session_input_optional, 1},
+    {"_onnxr_onnx_session_input_shapes",   (DL_FUNC) &_onnxr_onnx_session_input_shapes,   1},
+    {"_onnxr_onnx_session_input_types",    (DL_FUNC) &_onnxr_onnx_session_input_types,    1},
+    {"_onnxr_onnx_session_output_count",   (DL_FUNC) &_onnxr_onnx_session_output_count,   1},
+    {"_onnxr_onnx_session_output_names",   (DL_FUNC) &_onnxr_onnx_session_output_names,   1},
+    {"_onnxr_onnx_session_output_shapes",  (DL_FUNC) &_onnxr_onnx_session_output_shapes,  1},
+    {"_onnxr_onnx_session_output_types",   (DL_FUNC) &_onnxr_onnx_session_output_types,   1},
+    {"_onnxr_onnx_unload_lib",             (DL_FUNC) &_onnxr_onnx_unload_lib,             0},
+    {"_onnxr_onnx_version",                (DL_FUNC) &_onnxr_onnx_version,                0},
     {NULL, NULL, 0}
 };
 }